requirements.txt
# Optional dependencies for providers and tests
openai>=1.52.0
requests>=2.32.0
python-dotenv>=1.0.1
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pytest>=8.3.3
